Feature added for deleting images and product

+++ schema/resolvers/product.resolver.ts
@@ -5,11 +5,16 @@ import Product from "../../dataLayer/schema/Product";
 import verifyUser from "../../utilities/verifyUser";
 import uploadImage from "../../utilities/uploadImage";
 import { inputProductImg } from '../../typeDefs';
+import deleteImage from '../../utilities/deleteImage';
 
 const productResolver = {
   Query: {
     products: async (parent: any, args: any, context: any) => {
-      const admin = context.admin
+      if (!context.token) {
+        throw new Error('Not Authenticated')
+      }
+
+      const admin = verifyUser(context.token)
       if (!admin) {
         throw new Error('Not Authenticated')
       }
@@ -40,53 +45,63 @@ const productResolver = {
 
       const { title, slug, description, colors, sizes, price, category, quantity, sku, stockStatus, featured, imgs } = args.input
 
-      
-      
+      console.log(typeof price);
+
+
       const productExists = await Product.findOne({ slug: slug.toLowerCase() })
       if (productExists) {
         throw new Error('Product Already Exists')
       }
 
-      const directory = path.join(process.cwd(), `public/upload/product`)
+      const folder = `public/upload/product`
 
-
-      // Check if the directory exists
-      if (!fs.existsSync(directory)) {
-        // Create the directory (and any necessary subdirectories)
-        fs.mkdirSync(directory, { recursive: true });
-      }
 
       let newImgs = []
 
       try {
-        const uploadPromises = imgs.map((item: inputProductImg) => uploadImage(item, directory))
-         newImgs = await Promise.all(uploadPromises);
+        const uploadPromises = imgs.map((item: inputProductImg) => uploadImage(item, folder))
+        newImgs = await Promise.all(uploadPromises);
       } catch (error) {
         throw error
       }
 
       const newProduct = new Product({
-        title, slug, description, colors, sizes, price, quantity, category, sku, stockStatus, featured, imgs:newImgs
+        title, slug, description, colors, sizes, price, quantity, category, sku, stockStatus, featured, imgs: newImgs
       })
 
       return await newProduct.save()
     },
 
-
-
     // Delete Product Mutation
-    deleteProduct: async (parent: any, args: any) => {
+    deleteProduct: async (parent: any, args: any, context: any) => {
+
+      if (!context.token) {
+        throw new Error('Not Authenticated')
+      }
+
+      const admin = verifyUser(context.token)
+      if (!admin) {
+        throw new Error('Not Authenticated')
+      }
+
       const { id } = args
 
       try {
-        const productExists = await Product.findByIdAndDelete(id)
-        if (productExists) {
-          return {
-            success: true,
-          }
+        const product = await Product.findById(id)
+        if (product) {
+          const imgLinks = product?.imgs.map(img => img.url)
 
+          imgLinks.map(link => deleteImage(link))
+
+          const productExists = await Product.findByIdAndDelete(id)
+          if (productExists) {
+            return {
+              success: true,
+            }
+          }
         }
-        throw new Error('User not found')
+
+        throw new Error('Bad Request')
       } catch (error) {
         if (error instanceof Error) {
           throw error
